Fold pack-wrapper into pack
